Refactored useMetadata (again)
