+++ frontend/src/Components/Metadata/ExpiryMetadataType.tsx
@@ -1,5 +1,5 @@
-import React from "react";
-import { metadataComponentProps } from '../metadataTypesReducer';
+import React from "react"
+import { MetadataComponentProps, MetadataType } from '../../Utils/Metadata/combinedReducer'
 
 type Command =
     | { path: "expiry", action: "INIT" }
@@ -13,15 +13,14 @@ type State = {
     validUntil: string
 }
 
-const empty: State = {
-    validFrom: "",
-    validUntil: ""
-}
 
 const reducer = (state: State, command: Command): State => {
     switch (command.action) {
         case "INIT":
-            return empty
+            return {
+                validFrom: "",
+                validUntil: ""
+            }
             
         case "METADATA_LOADED":
             return {
@@ -44,8 +43,8 @@ const reducer = (state: State, command: Command): State => {
     }
 }
 
-const Component = (props: metadataComponentProps) => {
-    const state = props.state as State
+const Component = (props: MetadataComponentProps) => {
+    const state = props.state
     const dispatch = props.dispatch as React.Dispatch<Command>
 
     return (
@@ -73,9 +72,8 @@ const Component = (props: metadataComponentProps) => {
 }
 
 export default {
-    name: "Expiry",
+    displayName: "Expiry",
     path: "expiry",
-    empty,
     component: Component,
     reducer
-} as const
+} as MetadataType
