+++ frontend/src/Components/Metadata/index.ts
@@ -1,0 +1,11 @@
+import { MetadataType } from '../../Utils/Metadata/combinedReducer'
+import ExpiryMetadataComponent from './ExpiryMetadataType'
+import ReceiptMetadataComponent from './ReceiptMetadataType'
+
+
+export const allMetadataTypes: Array<MetadataType> = [
+    ReceiptMetadataComponent,
+    ExpiryMetadataComponent
+]
+
+
